UPDATE/validate max size value

+++ src/main/java/com/capstone/ads/dto/customer_choice_size/CustomerChoicesSizeCreateRequest.java
@@ -11,5 +11,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class CustomerChoicesSizeCreateRequest {
     @Positive(message = "Size Value must greater than 0")
-    Double sizeValue;
+    Float sizeValue;
 }
